+++ BNA_KC_Weapons/Config.cpp
@@ -20,188 +20,6 @@ class CfgWeapons
     // │      Optics      │
     // └──────────────────┘
     #include "Optics.hpp"
-	
-    // ┌───────────────────┐
-    // │      Weapons      │
-    // └───────────────────┘
-    class WeaponSlotsInfo;
-    class Default {};
-    class RifleCore: Default {};
-    class Rifle: RifleCore
-    {
-        class WeaponSlotsInfo {};
-    };
-    class Rifle_Base_F: Rifle {};
-    class 3AS_WestarM5_Base_F: Rifle_Base_F
-    {
-        // Attachments
-        class WeaponSlotsInfo;
-        // Fire Modes
-        class Single;
-        class Burst;
-        class FullAuto;
-        class 3AS_WestarM5_GL_F;
-    };
-    class 3AS_WestarM5_F: 3AS_WestarM5_Base_F
-    {
-        class WeaponSlotsInfo: WeaponSlotsInfo
-        {
-            // Attachments
-            class CowsSlot;
-            class PointerSlot;
-        };
-    };
-    class BNA_KC_WestarM5: 3AS_WestarM5_F
-    {
-        // Mod Info
-        dlc = "BNA_KC";
-        author = "SweMonkey and DartRuffian";
-
-        // Scope
-        scope = 2;
-        scopeArsenal = 2;
-
-        displayName = "[KC] Westar-M5";
-        picture = "\SWLW_clones\smgs\westar_m5\data\ui\WestarM5_ui.paa";
-		recoil = "recoil_smg_03";
-        canShootInWater = 1;
-
-        magazines[] = { "12thFleet_Mag_WestarM5" };
-        magazineWell[] = {};
-
-        class WeaponSlotsInfo: WeaponSlotsInfo
-        {
-            class CowsSlot: CowsSlot
-            {
-                compatibleItems[] = 
-                {
-                    // Short
-                    "Aux501_cows_RCO",
-                    "Aux501_cows_RCO_2",
-                    "Aux501_cows_RCO_3",
-                    // Medium
-                    "Aux501_cows_MRCO",
-                    "Aux501_cows_MRCO_2",
-                    "Aux501_cows_MRCO_3",
-                    // Long
-                    "Aux501_cows_DMS",
-                    "Aux501_cows_DMS_2",
-                    "Aux501_cows_DMS_3",
-                    "Aux501_cows_DMS_4",
-                    "Aux501_cows_HoloScope",
-                    "Aux501_cows_HoloScope_2",
-                    "Aux501_cows_HoloScope_3",
-                    // Sights
-                    "BNA_KC_WestarM5_Scope",
-                    "Aux501_cows_reflex_optic",
-                    "Aux501_cows_Holosight",
-                    "Aux501_cows_Holosight_2",
-                    "Aux501_cows_Holosight_3"
-                };
-            };
-        };
-    };
-
-    class 3AS_WestarM5_GL: 3AS_WestarM5_Base_F
-    {
-        class WeaponSlotsInfo: WeaponSlotsInfo
-        {
-            // Attachments
-            class CowsSlot;
-            class PointerSlot;
-        };
-        class Single: Single { class StandardSound; };
-        class Burst: Burst { class StandardSound; };
-        class FullAuto: FullAuto { class StandardSound; };
-        class 3AS_WestarM5_GL_F;
-    };
-    class BNA_KC_WestarM5_UGL: 3AS_WestarM5_GL
-    {
-        // Mod Info
-        dlc = "BNA_KC";
-        author = "SweMonkey and DartRuffian";
-
-        // Scope
-        scope = 2;
-        scopeArsenal = 2;
-
-        displayName = "[KC] Westar-M5 UGL";
-        picture = "\SWLW_clones\smgs\westar_m5\data\ui\WestarM5_ui.paa";
-		recoil = "recoil_smg_03";
-        canShootInWater = 1;
-
-        magazines[] = { "12thFleet_Mag_WestarM5" };
-        magazineWell[] = {};
-
-        class WeaponSlotsInfo: WeaponSlotsInfo
-        {
-            class CowsSlot: CowsSlot
-            {
-                compatibleItems[] = 
-                {
-                    // Short
-                    "Aux501_cows_RCO",
-                    "Aux501_cows_RCO_2",
-                    "Aux501_cows_RCO_3",
-                    // Medium
-                    "Aux501_cows_MRCO",
-                    "Aux501_cows_MRCO_2",
-                    "Aux501_cows_MRCO_3",
-                    // Long
-                    "Aux501_cows_DMS",
-                    "Aux501_cows_DMS_2",
-                    "Aux501_cows_DMS_3",
-                    "Aux501_cows_DMS_4",
-                    "Aux501_cows_HoloScope",
-                    "Aux501_cows_HoloScope_2",
-                    "Aux501_cows_HoloScope_3",
-                    // Sights
-                    "BNA_KC_WestarM5_Scope",
-                    "Aux501_cows_reflex_optic",
-                    "Aux501_cows_Holosight",
-                    "Aux501_cows_Holosight_2",
-                    "Aux501_cows_Holosight_3"
-                };
-            };
-        };
-
-        // class Single: Single
-        // {
-        //     class StandardSound: StandardSound
-        //     {
-        //         begin1[] = {"BNA_KC_Gear\Weapons\Data\Audio\BNA_KC_Westar_Fire1", 1, 1, 1800};
-        //         begin2[] = {"BNA_KC_Gear\Weapons\Data\Audio\BNA_KC_Westar_Fire2", 1, 1, 1800};
-        //         begin3[] = {"BNA_KC_Gear\Weapons\Data\Audio\BNA_KC_Westar_Fire3", 1, 1, 1800};
-        //         begin4[] = {"BNA_KC_Gear\Weapons\Data\Audio\BNA_KC_Westar_Fire4", 1, 1, 1800};
-        //         soundBegin[] = {begin1, 0.25, begin2, 0.25, begin3, 0.25, begin4, 0.25};
-        //     };
-        // };
-
-        class 3AS_WestarM5_GL_F: 3AS_WestarM5_GL_F
-        {
-            displayName = "Grenade Launcher";
-            canShootInWater = 0;
-            magazines[] =
-            {
-                "3Rnd_UGL_FlareWhite_F",
-                "3Rnd_UGL_FlareRed_F",
-                "BNA_KC_3Rnd_UGL_FlareBlue",
-                "3Rnd_UGL_FlareCIR_F",
-                "3Rnd_Smoke_Grenade_shell",
-                "3Rnd_SmokeRed_Grenade_shell",
-                "3Rnd_SmokeGreen_Grenade_shell",
-                "3Rnd_SmokeYellow_Grenade_shell",
-                "3Rnd_SmokePurple_Grenade_shell",
-                "3Rnd_SmokeBlue_Grenade_shell",
-                "3Rnd_SmokeOrange_Grenade_shell",
-                "3Rnd_HE_Grenade_shell",
-                "ACE_HuntIR_M203"
-            };
-            magazineWell[] = {}; // Removes extra magazines
-        };
-    };
-
-
     // ┌────────────────────┐
     // │      Grenades      │
     // └────────────────────┘
